Add NSGA-II examples

+++ jmetal/util/observer.py
@@ -190,7 +190,7 @@ class VisualizerObserver(Observer):
     def __init__(self,
                  reference_front: List[S] = None,
                  reference_point: list = None,
-                 display_frequency: float = 1.0) -> None:
+                 display_frequency: int = 1) -> None:
         self.figure = None
         self.display_frequency = display_frequency
